PropositionDeleteView in proposition.views updated

+++ proposition/views/delete_proposition_view.py
@@ -7,6 +7,7 @@ from django.shortcuts import redirect, render
 
 from proposition.management.engine.manager import Manager
 from proposition.models.proposition import Proposition
+from proposition.models.status import Status
 
 
 class DeletePropositionView(LoginRequiredMixin, View):
@@ -23,10 +24,13 @@ class DeletePropositionView(LoginRequiredMixin, View):
         self.context = {
             'proposition': None
         }
+        self.msg_post_success = "Suppression réussie"
+        self.mgs_post_no_delete = (
+            "Une proposition avec ce satut ne peut pas être supprimée"
+        )
         self.msg_not_owner = (
             "Le créateur seulement peut supprimer la proposition"
         )
-        self.msg_post_success = "Suppression de proposition réussie"
     
     def get(self, request, id_proposition):
         """Delete proposition view method on client get request.
@@ -46,13 +50,22 @@ class DeletePropositionView(LoginRequiredMixin, View):
         """
         proposition = Proposition.objects.get(pk=id_proposition)
         if proposition.proposition_creator_id == request.user.id:
-            proposition.delete()
-            messages.add_message(
-                request, messages.SUCCESS, self.msg_post_success
-            )
-            return redirect(self.alternative_one_view_name)
+            if proposition.proposition_status.name in ['Terminé', 'Annulé']:
+                messages.add_message(
+                    request, messages.WARNING, self.mgs_post_no_delete
+                )
+            elif proposition.proposition_status.name == 'Nouveau':
+                proposition.delete()
+                messages.add_message(
+                    request, messages.SUCCESS, self.msg_post_success
+                )
+            else:
+                status = Status.objects.get(name__exact='Annulé')
+                proposition.proposition_status = status
+                proposition.save()
+                messages.add_message(
+                    request, messages.SUCCESS, self.msg_post_success
+                )
         else:
-            messages.add_message(
-                request, messages.ERROR, self.msg_not_owner
-            )
-            return redirect(self.alternative_one_view_name)
+            messages.add_message(request, messages.ERROR, self.msg_not_owner)
+        return redirect(self.alternative_one_view_name)
